requirements.txt
attrs==17.4.0
bleach==2.1.1
boto==2.48.0
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
Django==1.11
djangorestframework==3.7.1
entrypoints==0.2.3
ftfy==4.4.3
future==0.16.0
gensim==3.1.0
googletrans==2.2.0
gunicorn==19.7.1
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
langid==1.1.6
MarkupSafe==1.0
matplotlib==2.1.0
mediacloud-cliff==2.0.2
mediameter-cliff==1.4.0
mistune==0.8.1
msgpack-numpy==0.4.1
msgpack-python==0.4.8
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.2.1
numpy==1.13.3
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pathlib==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.4.1
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
rake-nltk==1.0.1
regex==2017.4.5
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0rc1
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.3
spacy==2.0.5
stop-words==2015.2.23.1
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
thinc==6.10.2
toolz==0.8.2
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.6
wrapt==1.10.11
